requirements.txt
Flask==2.0.2
google-cloud-vision==2.6.2
Pillow==8.4.0